[CDEC-346] Refactor Pos.Core.Common module set
