+++ core/src/Pos/Core/Common.hs
@@ -1,11 +1,37 @@
 module Pos.Core.Common
-       ( module Pos.Core.Common.Address
+       ( module Pos.Core.Common.AddrAttributes
+       , module Pos.Core.Common.Address
+       , module Pos.Core.Common.AddressHash
+       , module Pos.Core.Common.AddrSpendingData
+       , module Pos.Core.Common.AddrStakeDistribution
+       , module Pos.Core.Common.BlockCount
+       , module Pos.Core.Common.ChainDifficulty
+       , module Pos.Core.Common.Coeff
        , module Pos.Core.Common.Coin
-       , module Pos.Core.Common.Fee
-       , module Pos.Core.Common.Types
+       , module Pos.Core.Common.CoinPortion
+       , module Pos.Core.Common.Script
+       , module Pos.Core.Common.SharedSeed
+       , module Pos.Core.Common.SlotLeaders
+       , module Pos.Core.Common.StakeholderId
+       , module Pos.Core.Common.Stakes
+       , module Pos.Core.Common.TxFeePolicy
+       , module Pos.Core.Common.TxSizeLinear
        ) where
 
+import           Pos.Core.Common.AddrAttributes
 import           Pos.Core.Common.Address
+import           Pos.Core.Common.AddressHash
+import           Pos.Core.Common.AddrSpendingData
+import           Pos.Core.Common.AddrStakeDistribution
+import           Pos.Core.Common.BlockCount
+import           Pos.Core.Common.ChainDifficulty
+import           Pos.Core.Common.Coeff
 import           Pos.Core.Common.Coin
-import           Pos.Core.Common.Fee
-import           Pos.Core.Common.Types
+import           Pos.Core.Common.CoinPortion
+import           Pos.Core.Common.Script
+import           Pos.Core.Common.SharedSeed
+import           Pos.Core.Common.SlotLeaders
+import           Pos.Core.Common.StakeholderId
+import           Pos.Core.Common.Stakes
+import           Pos.Core.Common.TxFeePolicy
+import           Pos.Core.Common.TxSizeLinear
