+++ core/src/Pos/Core/Common/AddrAttributes.hs
@@ -1,0 +1,91 @@
+module Pos.Core.Common.AddrAttributes
+       ( AddrAttributes (..)
+       ) where
+
+import           Universum
+
+import qualified Data.ByteString.Lazy as LBS
+import qualified Data.Text.Buildable as Buildable
+import           Formatting (bprint, build, builder, (%))
+
+import           Pos.Binary.Class (Bi, decode, encode)
+import qualified Pos.Binary.Class as Bi
+import           Pos.Crypto.HD (HDAddressPayload)
+import           Pos.Data.Attributes (Attributes (..), decodeAttributes, encodeAttributes)
+
+import           Pos.Core.Common.AddrStakeDistribution
+
+-- | Additional information stored along with address. It's intended
+-- to be put into 'Attributes' data type to make it extensible with
+-- softfork.
+data AddrAttributes = AddrAttributes
+    { aaPkDerivationPath  :: !(Maybe HDAddressPayload)
+    , aaStakeDistribution :: !AddrStakeDistribution
+    } deriving (Eq, Ord, Show, Generic, Typeable)
+
+instance Buildable AddrAttributes where
+    build (AddrAttributes {..}) =
+        bprint
+            ("AddrAttributes { stake distribution: "%build%
+             ", derivation path: "%builder%" }")
+            aaStakeDistribution
+            derivationPathBuilder
+      where
+        derivationPathBuilder =
+            case aaPkDerivationPath of
+                Nothing -> "{}"
+                Just _  -> "{path is encrypted}"
+
+instance NFData AddrAttributes
+
+{- NOTE: Address attributes serialization
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+'Attributes' are conceptually a map, where keys are numbers ('Word8').
+
+For address there are two attributes:
+• 0 — stake distribution, defaults to 'BootstrapEraDistr';
+• 1 — derivation path, defaults to 'Nothing'.
+
+-}
+
+instance Bi (Attributes AddrAttributes) where
+    -- FIXME @avieth it was observed that for a 150kb block, this call to
+    -- encodeAttributes allocated 3.685mb
+    -- Try using serialize rather than serialize', to avoid the
+    -- toStrict call.
+    -- Also consider using a custom builder strategy; serialized attributes are
+    -- probably small, right?
+    encode attrs@(Attributes {attrData = AddrAttributes derivationPath stakeDistr}) =
+        encodeAttributes listWithIndices attrs
+      where
+        listWithIndices :: [(Word8, AddrAttributes -> LBS.ByteString)]
+        listWithIndices =
+            stakeDistributionListWithIndices <> derivationPathListWithIndices
+        stakeDistributionListWithIndices =
+            case stakeDistr of
+                BootstrapEraDistr -> []
+                _                 -> [(0, Bi.serialize . aaStakeDistribution)]
+        derivationPathListWithIndices =
+            case derivationPath of
+                Nothing -> []
+                -- 'unsafeFromJust' is safe, because 'case' ensures
+                -- that derivation path is 'Just'.
+                Just _ ->
+                    [(1, Bi.serialize . unsafeFromJust . aaPkDerivationPath)]
+        unsafeFromJust =
+            fromMaybe
+                (error "Maybe was Nothing in Bi (Attributes AddrAttributes)")
+
+    decode = decodeAttributes initValue go
+      where
+        initValue =
+            AddrAttributes
+            { aaPkDerivationPath = Nothing
+            , aaStakeDistribution = BootstrapEraDistr
+            }
+        go n v acc =
+            case n of
+                0 -> (\distr -> Just $ acc {aaStakeDistribution = distr }    ) <$> Bi.deserialize v
+                1 -> (\deriv -> Just $ acc {aaPkDerivationPath = Just deriv }) <$> Bi.deserialize v
+                _ -> pure Nothing
