+++ core/src/Pos/Core/Common/StakeholderId.hs
@@ -1,0 +1,10 @@
+module Pos.Core.Common.StakeholderId
+       ( StakeholderId
+       ) where
+
+import           Pos.Crypto.Signing (PublicKey)
+
+import           Pos.Core.Common.AddressHash
+
+-- | Stakeholder identifier (stakeholders are identified by their public keys)
+type StakeholderId = AddressHash PublicKey
